+++ R/argonButton.R
@@ -2,6 +2,7 @@
 #'
 #' Build an argon button
 #'
+#' @param src Button external link.
 #' @param name Button label.
 #' @param status Button color. See \url{https://demos.creative-tim.com/argon-design-system/docs/components/buttons.html}.
 #' @param icon Button icon.
@@ -11,7 +12,6 @@
 #' @param outline Whether to outline the button. FALSE by default.
 #' @param toggle_modal Whether to use th button for a modal. FALSE by default.
 #' @param modal_id If toggle_modal is TRUE, nedd to provide the modal targeted.
-#' @param rounded Whether to set the rounded style. FALSE by default.
 #'
 #' @examples
 #' if(interactive()){
@@ -29,16 +29,14 @@
 #' @author David Granjon, \email{dgranjon@@ymail.com}
 #'
 #' @export
-argonButton <- function(name = NULL, status = "default", icon = NULL, 
+argonButton <- function(src = NULL, name = NULL, status = "default", icon = NULL, 
                         size = NULL, block = FALSE, disabled = FALSE, 
-                        outline = FALSE, toggle_modal = FALSE, modal_id = NULL,
-                        rounded = FALSE) {
+                        outline = FALSE, toggle_modal = FALSE, modal_id = NULL) {
   
   btnCl <- "btn"
   if (!is.null(status)) btnCl <- paste0(btnCl, " btn-", if (outline) "outline-" else NULL, status)
   if (!is.null(size)) btnCl <- paste0(btnCl, " btn-", size)
   if (block) btnCl <- paste0(btnCl, " btn-block")
-  if (rounded) btnCl <- paste0(btnCl, " rounded-circle")
   
   htmltools::tags$button(
     class = btnCl,
@@ -48,15 +46,62 @@ argonButton <- function(name = NULL, status = "default", icon = NULL,
     `data-target` = if (toggle_modal) paste0("#", modal_id) else NULL,
     if (!is.null(icon)) {
       if (!is.null(name)) {
-        htmltools::tagList(
+        htmltools::tags$a(
+          href = src,
+          target = "_blank",
           htmltools::tags$span(class = "btn-inner--icon", argonIcon(icon)),
           htmltools::tags$span(class = "btn-inner--text", name)
         )
       } else {
-        htmltools::tags$span(class = "btn-inner--icon", argonIcon(icon))
+        htmltools::tags$a(
+          href = src,
+          target = "_blank",
+          htmltools::tags$span(class = "btn-inner--icon", argonIcon(icon))
+        )
       }
     } else {
-      name
+      htmltools::tags$a(
+        href = src,
+        target = "_blank",
+        name
+      )
     }
   )
+}
+
+
+
+
+#' Create a Boostrap 4 argon social button
+#'
+#' Build an argon social button
+#'
+#' @param src Button external link.
+#' @param status Button color. See \url{https://demos.creative-tim.com/argon-design-system/docs/components/buttons.html}.
+#' @param icon Button icon.
+#'
+#' @examples
+#' if(interactive()){
+#'  library(argonR)
+#'  argonSocialButton(
+#'   src = "http://rinterface.com",
+#'   status = "danger",
+#'   icon = "atom"
+#'  )
+#' }
+#'
+#' @author David Granjon, \email{dgranjon@@ymail.com}
+#'
+#' @export
+argonSocialButton <- function(src = NULL, status = "default", icon = NULL) {
+  btnCl <- "btn btn-icon-only rounded-circle"
+  if (!is.null(status)) btnCl <- paste0(btnCl, " btn-", status)
+  
+  htmltools::tags$a(
+    class = btnCl,
+    href = src,
+    target = "_blank",
+    htmltools::tags$i(class = "fa fa-", icon)
+  )
+  
 }